+++ library/src/main/scala/minimal_algorithms/aggregation_function/AggregationFunction.scala
@@ -5,12 +5,12 @@ abstract class AggregationFunction extends Serializable {
     * @return Default value for given aggregation function. It is a start value for iteration ex. foldLeft(defaultValue){...}
     *         and as well it is used in RangeTree initialization.
     */
-  def defaultValue: Int
+  def defaultValue: Double
 
   /**
     * @return Aggregation function ex. (x: Int, y: Int) => x + y
     */
-  def apply: (Int, Int) => Int
+  def apply: (Double, Double) => Double
 
   /**
     * @return Flag indicating whether we should use average result or not.
