+++ library/src/main/scala/minimal_algorithms/aggregation_function/MaxAggregation.scala
@@ -1,7 +1,7 @@
 package minimal_algorithms.aggregation_function
 
 class MaxAggregation extends AggregationFunction {
-  override def defaultValue: Int = Int.MinValue
-  override def apply: (Int, Int) => Int = (x: Int, y: Int) => math.max(x, y)
+  override def defaultValue: Double = Double.MinValue
+  override def apply: (Double, Double) => Double = (x: Double, y: Double) => math.max(x, y)
   override def average: Boolean = false
 }
