+++ library/src/main/scala/minimal_algorithms/aggregation_function/MinAggregation.scala
@@ -1,7 +1,7 @@
 package minimal_algorithms.aggregation_function
 
 class MinAggregation extends AggregationFunction {
-  override def defaultValue: Int = Int.MaxValue
-  override def apply: (Int, Int) => Int = (x: Int, y: Int) => math.min(x, y)
+  override def defaultValue: Double = Double.MaxValue
+  override def apply: (Double, Double) => Double = (x: Double, y: Double) => math.min(x, y)
   override def average: Boolean = false
 }
