+++ library/src/main/scala/minimal_algorithms/semi_join/ExampleSemiJoin.scala
@@ -15,12 +15,12 @@ object ExampleSemiJoin {
     val inputR = spark.sparkContext.textFile(inputPathR)
     val inputMappedR = inputR.map(line => {
       val p = line.split(' ')
-      new SemiJoinType(p(0).toInt, p(1).toInt, SemiJoinTypeEnum.RType)})
+      new SemiJoinType(p(0).toInt, p(1).toDouble, SemiJoinTypeEnum.RType)})
 
     val inputT = spark.sparkContext.textFile(inputPathT)
     val inputMappedT = inputT.map(line => {
       val p = line.split(' ')
-      new SemiJoinType(p(0).toInt, p(1).toInt, SemiJoinTypeEnum.TType)})
+      new SemiJoinType(p(0).toInt, p(1).toDouble, SemiJoinTypeEnum.TType)})
 
     val minimalSemiJoin = new MinimalSemiJoin(spark, numOfPartitions).importObjects(inputMappedR, inputMappedT)
     minimalSemiJoin.execute.map(res => res.getKey.toString + " " + res.getWeight.toString).saveAsTextFile(outputPath)
